Created initial version of the max-login-attempts-spring-boot-starter
This is the initial version of the starter,
it can be used to put a maximum attempt on
the number of login failures a user can make
from a specific IP-address. The use-case of
this is to prevent hackers from brute-forcing
an account.

After the maximum amount of failed
attempts, the user is blocked for a certain
amount of time. During this time, an error
will be presented when the user attempts
to login, also if the credentials are
correct. The waiting time and amount of tries
are all configurable.