Simplified the LoginAttemptServiceImplementation
- Added the Attempt class for holding the attempts and unblock data.
- Moved the default clear all cron to the LoginAttemptConfiguration class,
so that the property can be autocompleted in an IDEA.
- Added the spring-boot-configuration-processor library so that the
configuration properties can be autocompleted in supported IDEA's of the
library.
- In LoginAttemptServiceImplementation merged the attemptsCache and
blockedUsernameIPAddresses to the userCache for storing the new Attempt
class.
- Property cooldown-in-ms was incorrecly documented, because the property
was cooldown. Fixed this by changing the property in
LoginAttemptConfiguration from cooldown to cooldownInMs so the value is
correctly mapped